outbound: Split sidecar and ingress stack modules
In preparation for new target types, this change splits the ingress and
proxy stack builders into ingress and sidecar modules, respectively.
